Pull BlockSupplier.threadSafe() implementation into AbstractBlockSupplier
