+++ src/main/java/net/imglib2/algorithm/blocks/PrimitiveBlocksSupplier.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -33,17 +33,12 @@
  */
 package net.imglib2.algorithm.blocks;
 
-import java.util.function.Supplier;
-
 import net.imglib2.blocks.PrimitiveBlocks;
 import net.imglib2.type.NativeType;
-import net.imglib2.util.CloseableThreadLocal;
 
-class PrimitiveBlocksSupplier< T extends NativeType< T > > implements BlockSupplier< T >
+class PrimitiveBlocksSupplier< T extends NativeType< T > > extends AbstractBlockSupplier< T >
 {
 	private final PrimitiveBlocks< T > blocks;
-
-	private Supplier< BlockSupplier< T > > threadSafeSupplier;
 
 	PrimitiveBlocksSupplier( final PrimitiveBlocks< T > blocks )
 	{
@@ -79,44 +74,5 @@ class PrimitiveBlocksSupplier< T extends NativeType< T > > implements BlockSuppl
 	{
 		final PrimitiveBlocks< T > blocksCopy = blocks.independentCopy();
 		return blocksCopy == blocks ? this : new PrimitiveBlocksSupplier<>( blocksCopy );
-	}
-
-	@Override
-	public BlockSupplier< T > threadSafe()
-	{
-		if ( threadSafeSupplier == null )
-			threadSafeSupplier = CloseableThreadLocal.withInitial( this::independentCopy )::get;
-		return new BlockSupplier< T >()
-		{
-			@Override
-			public T getType()
-			{
-				return blocks.getType();
-			}
-
-			@Override
-			public int numDimensions()
-			{
-				return blocks.numDimensions();
-			}
-
-			@Override
-			public void copy( final long[] srcPos, final Object dest, final int[] size )
-			{
-				threadSafeSupplier.get().copy( srcPos, dest, size );
-			}
-
-			@Override
-			public BlockSupplier< T > independentCopy()
-			{
-				return PrimitiveBlocksSupplier.this.independentCopy().threadSafe();
-			}
-
-			@Override
-			public BlockSupplier< T > threadSafe()
-			{
-				return this;
-			}
-		};
 	}
 }
